% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractR.R
\name{extractR}
\alias{extractR}
\title{Extracts pixel values from Landsat imagery}
\usage{
extractR(wkdir, imdir, shp, shp.ID, option)
}
\arguments{
\item{wkdir}{character string of the working directory where your original 
shape file is located}

\item{imdir}{character string of the directory where imagery is located to 
path/row level}

\item{shp}{character string of the name of the shape file (no extension)}

\item{shp.ID}{character string of the name of the field in the attribute
table of the shape file that contains the unique entries (e.g.site numbers)}

\item{option}{character string choice for extracted values. One of "i35", 
"ndvi", "b1", "b2", "b3", "b4", "b5", "b6"}
}
\value{
The csv from step 5 will contain all extracted data (pixel values or
VI) for each site. An 'NA' in the data indicates that a site was impacted by 
cloud or fell within a missing data stripe (Landsat 7 post 2003).
}
\description{
This function must be run only after jpegR and visual QA of jpegs has been 
completed. It will extract either the pixel value or mean pixel values based 
on a point or polygon shape file respectively, from a stack of Landsat 
imagery. It can also extract some predefined vegetation indices (VI). This 
function:
\enumerate{
   \item copies individual site shape files to respective QA jpeg folders
   \item obtains the dates of the remaining jpegs found in each of the QA 
   jpeg folders
   \item uses the dates obtained from step 2 to extract data (pixel values or 
   VI) from the local USGS Landsat archive
   \item writes the results of step 3 to a csv file within each QA jpeg 
   folder
   \item combines all individual site data extract csv files and writes to 
   one csv file at the working directory level
   }
}
\examples{
\dontrun{
extractR("../wkdir", "../imdir", "plot_locations", "Plot_ID", "i35")
}

}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}
}
\keyword{utilities}
