% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmaskcheckR.R
\name{fmaskcheckR}
\alias{fmaskcheckR}
\title{A function used to create a list of image date folders based on outputs from 
fmask}
\usage{
fmaskcheckR(wkdir, imdir, shp, shp.ID)
}
\arguments{
\item{wkdir}{character string of the working directory where your original 
shape file is located}

\item{imdir}{character string of the directory where imagery is located to 
path/row level}

\item{shp}{character string of the name of the shape file (no extension)}

\item{shp.ID}{character string of the name of the field in the attribute
table of the shape file that contains the unique entries (e.g.site numbers)}
}
\value{
creates a list of image date folders in RDA format used in the jpegR
function
}
\description{
This function is used internally in the jpegR function to create a list of 
image date folders to process. This function differs from checkR by querying
fmask outputs on a site by site basis to reject dates with cloud. It also:
\itemize{
    \item splits an ESRI shape file into multiple shape files based on unique 
    entries in a field of the attribute table
    \item stores the split shape files in a folder named \emph{QAshapes}
    \item performs checks to see that leap year image folders are named 
    correctly
    \item outputs results in \strong{DateSiteList} RData file
    }
}
\examples{
\dontrun{
checkR("../wkdir", "../imdir", "plot_locations", "Plot_ID")
}


}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}
}
\keyword{utilities}
