% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpegR.R
\name{jpegR}
\alias{jpegR}
\title{A function that produces small jpegs of satellite imagery}
\usage{
jpegR(wkdir, imdir, shp, shp.ID, buff, red, green, blue, fmask)
}
\arguments{
\item{wkdir}{character string of the working directory where your original 
shape file is located}

\item{imdir}{character string of the directory where imagery is located to 
path/row level}

\item{shp}{character string of the name of the shape file (no extension)}

\item{shp.ID}{character string of the name of the field in the attribute
table of the shape file that contains the unique entries (e.g.site numbers)}

\item{buff}{numeric representing how much to buffer out the site in metres. 
It is a way to control zoom level for jpegs}

\item{red}{numeric representing what satellite band to display in the red 
channel}

\item{green}{numeric representing what satellite band to display in the green 
channel}

\item{blue}{numeric representing what satellite band to display in the blue 
channel}

\item{fmask}{character string 'y' or 'n' representing whether fmask products
exist in each of the image date folders. Suggest 'n' as majority of our 
Landsat archive has not been processed by fmask}
}
\value{
creates named folders, based on site location and dates of images 
processed, containing small jpeg images for visual QA
}
\description{
This function makes small jpeg images centred on and showing site locations. 
This step is part of the cloud quality check process prior to extracting 
satellite bands or indices. Although designed to help the QA process the jpegs 
can be useful for quickly checking many image dates for change (e.g. fires,
vegetation removal, etc). When run the function:
\enumerate{
    \item obtains a list of imagery folders to access. Can be from one of 2 
    sources. Either from auto cloud QA'd output (fmaskcheckR) or from all 
    available imagery output (checkR). These functions are called internally
    \item obtains individual site shps from \emph{QAshapes} folder
    \item creates a small jpeg of the individual sites for each image date as 
    per step 1 for the purposes of visual QA
    \item jpegs are based on band combination from user input and buffered by 
    desired distance
    \item all outputs are placed in appropriately named folders per site in 
    the working directory
    }
}
\examples{
\dontrun{
jpegR("../wkdir", "../imdir", "plot_locations", "Plot_ID", 2000, 5, 4, 2, 'n')
}

}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}
}
\keyword{utilities}
