% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shpSplitR.R
\name{shpSplitR}
\alias{shpSplitR}
\title{A tool for splitting ESRI shape files}
\usage{
shpSplitR(shp, shp.ID)
}
\arguments{
\item{shp}{character string of the name of the shape file (no extension)}

\item{shp.ID}{character string of the name of the field in the attribute
table of the shape file that contains the unique entries (e.g.site numbers)}
}
\value{
creates multiple ESRI shape files each named after the unique entry
found in 'shp.ID'.
}
\description{
Takes an ESRI shape file and splits it into multiple shape files based on
unique entries in a field of the attribute table.
}
\examples{
\dontrun{
shpSplitR("plot_locations", "Plot_ID")
}


}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}
}
\keyword{utilities}
